Extract getMatchingAnchorSite() to undocumented.js
